This is a connection object to a FRAM database. Returns a list used throughout the rest of this package which carries meta data.

This is a connection object to a FRAM database. Returns a list used throughout the rest of this package which carries meta data.

Usage

connect_fram_db(
  db_path,
  enforce_type = c("full", "transfer"),
  read_only = FALSE,
  quiet = FALSE
)

Arguments

db_path

Path to a FRAM database.

enforce_type

Not used

read_only

Optional argument to flag this connection as read-only (if set to TRUE). If TRUE, framrsquared functions that modify the database will abort rather than run. Use as a safety feature when working with a database that must not be modified.

quiet

Logical. Optitional argument; when TRUE, silences success message and database summary.

Details

Additional details...

Examples

if (FALSE) fram_db <- connect_fram_db('<path>') # \dontrun{}